+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add components to subtract and also to negate points in-circuit [#842]
+
+### Changed 
+
+- Update dependency `dusk-bls12_381` to 0.14
+
 ## [0.20.2] - 2024-11-14
 
 ### Changed 
@@ -601,6 +609,7 @@ is necessary since `rkyv/validation` was required as a bound.
 - Proof system module.
 
 <!-- ISSUES -->
+[#842]: https://github.com/dusk-network/plonk/issues/842
 [#834]: https://github.com/dusk-network/plonk/issues/834
 [#831]: https://github.com/dusk-network/plonk/issues/831
 [#819]: https://github.com/dusk-network/plonk/issues/819